backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.8.3
pytesseract==0.3.10
Pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
aiohttp==3.9.1
redis[asyncio]==5.0.1
